Fixes some package names and some refactors
