+++ build.sbt
@@ -149,6 +149,13 @@ lazy val `prometheus-client` = project
   .settings(moduleName := "mu-rpc-prometheus-client")
   .settings(prometheusClientSettings)
 
+lazy val `prometheus` = project
+  .in(file("modules/metrics/prometheus"))
+  .dependsOn(`internal-core`)
+  .dependsOn(testing % "test->test")
+  .settings(moduleName := "mu-rpc-prometheus")
+  .settings(prometheusMetricsSettings)
+
 ////////////////////
 //// DROPWIZARD ////
 ////////////////////
@@ -389,6 +396,7 @@ lazy val allModules: Seq[ProjectReference] = Seq(
   `dropwizard-server`,
   `dropwizard-client`,
   `dropwizard`,
+  `prometheus`,
   testing,
   ssl,
   `idlgen-core`,
